1. I updated the dependecies
2. I made the script to work so that you don't need to put your js and css files in the python script directory.
You just type in the terminal the path to your directory.
The script will search afterwards the js and css files in a 'dev' and 'pub' directory depending on the execution mode.
In case the 'dev' and 'pub' directories are not present in the project directory, the script will ask you if you wish to create them for your.

3. The 'html' game file that has the path to your js and css files inserted in the head of the file, will be saved in a directory named 'tmp'.
In case the 'tmp' directory is not present in the project directory, the script will ask you if you wish to create it for you.